requirements.txt
streamlit>=1.32.0
groq>=0.9.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
pandas>=2.0.0
